# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: SSAServer
  description: SSAサービスのサーバーサイドプログラム
  version: "0.1"
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /group/{group_id}:
    get:
      tags:
      - SSAServer
      summary: Return_data_list SSAServer
      description: データのリストを取得する
      operationId: SSAServer#Return_data_list
      parameters:
      - name: group_id
        in: path
        description: Group ID
        required: true
        type: string
      - name: Return_data_list_Request_Body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SSAServerReturnDataListRequestBody'
          required:
          - user_id
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/SSAServerSsaResultResponseCollection'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/SSAServerReturn_data_list_Invalid_Request_Response_Body'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/SSAServerReturn_data_list_Invalid_Group_ID_Response_Body'
      schemes:
      - http
    post:
      tags:
      - SSAServer
      summary: Save_data SSAServer
      description: データをサーバーへ保存する
      operationId: SSAServer#Save_data
      parameters:
      - name: group_id
        in: path
        description: Group ID
        required: true
        type: string
      - name: Save_data_Request_Body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SSAServerSaveDataRequestBody'
          required:
          - user_id
          - data_name
          - Data
      responses:
        "200":
          description: OK response.
          schema:
            type: boolean
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/SSAServerSave_data_Invalid_Data_Response_Body'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/SSAServerSave_data_Invalid_Group_ID_Response_Body'
      schemes:
      - http
  /group/{group_id}/{data_type}:
    get:
      tags:
      - SSAServer
      summary: Pick_up_data SSAServer
      description: データをサーバーから取得する
      operationId: SSAServer#Pick_up_data
      parameters:
      - name: group_id
        in: path
        description: Group ID
        required: true
        type: string
      - name: data_type
        in: path
        description: Data type
        required: true
        type: string
      - name: Pick_up_data_Request_Body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SSAServerPickUpDataRequestBody'
          required:
          - user_id
          - data_name
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/SSAServerPick_up_data_Response_BodyData_extended_with_image'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/SSAServerPick_up_data_Invalid_data_name_Response_Body'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/SSAServerPick_up_data_Invalid_Group_ID_Response_Body'
      schemes:
      - http
...
  /users/{user_id}:
    post:
      tags:
      - SSAServer
      summary: Change_group SSAServer
      description: グループIDを変更する
      operationId: SSAServer#Change_group
      parameters:
      - name: user_id
        in: path
        description: User ID
        required: true
        type: integer
      - name: Change_group_Request_Body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SSAServerChangeGroupRequestBody'
          required:
          - password
          - group_id
      responses:
        "200":
          description: OK response.
          schema:
            type: boolean
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/SSAServerChange_group_Invalid_Request_Response_Body'
      schemes:
      - http
    delete:
      tags:
      - SSAServer
      summary: Delete_user SSAServer
      description: 既存ユーザーの消去
      operationId: SSAServer#Delete_user
      parameters:
      - name: user_id
        in: path
        description: User ID
        required: true
        type: integer
      - name: Delete_user_Request_Body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SSAServerDeleteUserRequestBody'
          required:
          - password
      responses:
        "200":
          description: OK response.
          schema:
            type: boolean
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/SSAServerDelete_user_Invalid_Request_Response_Body'
      schemes:
      - http
definitions:
...
  SSAServerChange_group_Invalid_Request_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Change_group_Invalid_Request_Response_Body result type (default view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
  SSAServerChangeGroupRequestBody:
    title: SSAServerChangeGroupRequestBody
    type: object
    properties:
      group_id:
        type: string
        description: Group ID
        example: group2
      password:
        type: string
        description: User password
        example: pass1234
    example:
      group_id: group2
      password: pass1234
    required:
    - password
    - group_id
  SSAServerDelete_user_Invalid_Request_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Delete_user_Invalid_Request_Response_Body result type (default view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
  SSAServerDeleteUserRequestBody:
    title: SSAServerDeleteUserRequestBody
    type: object
    properties:
      password:
        type: string
        description: User Password
        example: pass12345
    example:
      password: pass12345
    required:
    - password
...
  SSAServerPick_up_data_Invalid_Group_ID_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Pick_up_data_Invalid_Group_ID_Response_Body result type (default
      view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
...
  SSAServerPick_up_data_Invalid_data_name_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Pick_up_data_Invalid_data_name_Response_Body result type (default
      view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
  SSAServerPick_up_data_Response_BodyData_extended_with_image:
    title: 'Mediatype identifier: application/vnd.ssa.result; view=data_extended_with_image'
    type: object
    properties:
      Data:
        description: Data
        example: Ipsa repudiandae voluptas labore voluptatem.
      Image:
        description: Image data
        example: Sint et et nam.
      data_name:
        type: string
        description: Data name
        example: Diary_312_2019-03-02_12-07-35
      data_type:
        type: integer
        description: Data's name
        example: 1
        format: int64
      image_name:
        type: string
        description: image's name
        example: huge
      title:
        type: string
        description: Data title
        example: hoge
    description: Pick_up_data_Response_Body result type (data_extended_with_image
      view) (default view)
    example:
      Data: Non fugiat rerum.
      Image: Velit quo quae et voluptatem occaecati quia.
      data_name: Diary_312_2019-03-02_12-07-35
      data_type: 1
      image_name: huge
      title: hoge
    required:
    - Data
    - data_type
    - data_name
  SSAServerPickUpDataRequestBody:
    title: SSAServerPickUpDataRequestBody
    type: object
    properties:
      data_name:
        type: string
        description: Data name
        example: Record_12_2019-06-02_12-07-35
      user_id:
        type: integer
        description: User ID
        example: 65
        format: int64
    example:
      data_name: Record_12_2019-06-02_12-07-35
      user_id: 65
    required:
    - user_id
    - data_name
...
  SSAServerReturn_data_list_Invalid_Group_ID_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Return_data_list_Invalid_Group_ID_Response_Body result type (default
      view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
  SSAServerReturn_data_list_Invalid_Request_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Return_data_list_Invalid_Request_Response_Body result type (default
      view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
  SSAServerReturnDataListRequestBody:
    title: SSAServerReturnDataListRequestBody
    type: object
    properties:
      user_id:
        type: integer
        description: User ID
        example: 537829
        format: int64
    example:
      user_id: 537829
    required:
    - user_id
  SSAServerSave_data_Invalid_Data_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Save_data_Invalid_Data_Response_Body result type (default view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
  SSAServerSave_data_Invalid_Group_ID_Response_Body:
    title: 'Mediatype identifier: application/vnd.ssa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Return false when Error occur
        default: true
        example: true
      message:
        type: string
        description: Error returned.
        example: 不正なリクエストです
    description: Save_data_Invalid_Group_ID_Response_Body result type (default view)
    example:
      fault: true
      message: 不正なリクエストです
    required:
    - fault
    - message
...
  SSAServerSaveDataRequestBody:
    title: SSAServerSaveDataRequestBody
    type: object
    properties:
      Data:
        description: Data
        example: Quia expedita ut dolorem quia.
      Image:
        description: Image
        example: Molestiae ut sint sunt deserunt quia.
      data_name:
        type: string
        description: Data name
        example: Diary_312_2019-03-02_12-07-35
      data_type:
        type: integer
        description: Data name
        example: 1
        format: int64
      image_name:
        type: string
        description: Image name
        example: Image_2017-05-25-26-32
      title:
        type: string
        description: Diary title
        example: たいとる
      user_id:
        type: integer
        description: User ID
        example: 28532
        format: int64
    example:
      Data: Adipisci quasi sit.
      Image: Eaque laudantium aut eum eos harum ea.
      data_name: Diary_312_2019-03-02_12-07-35
      data_type: 1
      image_name: Image_2017-05-25-26-32
      title: たいとる
      user_id: 28532
    required:
    - user_id
    - data_name
    - Data
  SSAServerSsaResultResponseCollection:
    title: 'Mediatype identifier: application/vnd.ssa.result; type=collection; view=default'
    type: array
    items:
      $ref: '#/definitions/SsaResultResponse'
    description: Return_data_list_Response_Body is the result type for an array of
      SsaResultResponse (default view)
    example:
    - mail: hoge@hoge.com
      user_id: 234
      user_name: 平野 竜也
    - mail: hoge@hoge.com
      user_id: 234
      user_name: 平野 竜也
  SsaResultResponse:
    title: 'Mediatype identifier: application/vnd.ssa.result; view=default'
    type: object
    properties:
      mail:
        type: string
        description: User mail-address
        example: hoge@hoge.com
      user_id:
        type: integer
        description: User id
        example: 234
        format: int64
      user_name:
        type: string
        description: User Name
        example: 平野 竜也
    description: SsaResultResponse result type (default view)
    example:
      mail: hoge@hoge.com
      user_id: 234
      user_name: 平野 竜也